qwen3-tts/pyproject.toml
[project]
name = "qwen3-tts"
version = "0.1.0"
description = "TTS using NVIDIA MagpieTTS"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27",
    "soundfile>=0.13.1",
]

[project.optional-dependencies]
server = [
    "nemo_toolkit[tts] @ git+https://github.com/NVIDIA/NeMo.git",
    "kaldialign",
    "fastapi",
    "uvicorn[standard]",
]
